Resulst of debugging: combining scalability with client/server, with multi-container processes
examples/xml/bias.xml: revert to not be redundant with testbias2
os/linux/src/OcpiOsDebug.cxx: add range to time values
runtime/application/include/OcpiApplication.h: warning for size_t/unsigned
runtime/application/src/OcpiApplication.cxx: force proxies to the base container
runtime/application/src/RemoteServer.cxx: support initial connect info in launch message,
-- and pass connection options, and clear initial info when sent/used.
runtime/container/include/ContainerBasicPort.h: allow connectInProcess to set "done"
runtime/container/include/ContainerLauncher.h: add Transport struct initializer
runtime/container/src/Container.cxx: don't assume base container is container 0. look for first rcc.
-- also share singleton class (slightly enhanced) for local launcher
runtime/container/src/ContainerBasicPort.cxx: fix peek method, and too many other bugs...
-- mostly managing when connecting is truly "done", and clearing initial info correctly
runtime/container/src/ContainerLauncher.cxx: accomodate multiple collocated containers properly
-- and doing slaves before masters
-- and Transport constructor
runtime/container/src/ContainerLocalPort.cxx: remote connections to bridge ports.
runtime/container/src/ContainerManager.cxx: use singleton template for local launcher.
runtime/container/src/ContainerPort.cxx: remove duplicate (and bad) port destruction
runtime/dataplane/transport/impl/include/OcpiRDTInterface.h: more complete desc initialization
runtime/dataplane/transport/impl/include/OcpiTransportConstants.h: comment out unused values
runtime/dataplane/transport/impl/src/OcpiPort.cxx: descriptor constructor/initialization
runtime/dataplane/transport/impl/src/OcpiPortMetaData.cxx: more member initializers
runtime/ocl/src/OclContainer.cxx: ditto, for valgrind, plus leaks
runtime/remote/src/RemoteDriver.cxx: memory leak fix in return path
runtime/remote/src/RemoteLauncher.cxx: add connection options, broken "crew" attribute
runtime/util/driver/include/OcpiDriverManager.h: enhance singleton template
-- allow cleanup to test whether singleton exists.
runtime/util/ezxml/src/OcpiUtilEzxml.cxx: buffering bug when reads not fully satisfied
runtime/util/misc/src/OcpiUtilMisc.cxx: fix broken XML quote processing

+++ runtime/application/include/OcpiApplication.h
@@ -94,7 +94,7 @@ namespace OCPI {
 	// This array is of indices in the usedcontainer array, not global indices
 	unsigned m_usedContainer;       // when scale==1 this is it
 	unsigned *m_usedContainers;     // container for each crew member
-	unsigned m_firstMember;         // index of first member in launch members
+	size_t m_firstMember;           // index of first member in launch members
 	void collectCandidate(OCPI::Library::Candidate &c, unsigned n);
 	void finalizePortParam(OU::Assembly::Instance &ui, const OCPI::Util::PValue *params,
 			       const char *param);
